refactor(format): extract export rendering module
